Split _frame_helper into plain_text and noise
