+++ aioesphomeapi/_frame_helper/__init__.py
@@ -1,0 +1,7 @@
+from .noise import APINoiseFrameHelper
+from .plain_text import APIPlaintextFrameHelper
+
+__all__ = (
+    "APINoiseFrameHelper",
+    "APIPlaintextFrameHelper",
+)
